backend/requirements.txt
Django>=4.2
djangorestframework
pandas
python-dateutil
requests